Split pywandevent and clientapi into their own separate eggs.
This means that netevmon can depend upon them without requiring a full
NNTSC install.

It does mean that the names of various packages have had to change and
anything that was in lib but used by other programs has had to move into the
nntscclient egg. As a result, there may be a few broken imports / dependencies
but I'll hopefully test all of that out soon.